+++ helpers.go
@@ -12,6 +12,11 @@
 
 package tools4clj
 
+import (
+	"bufio"
+	"os"
+)
+
 func join(items []string, char string) string {
 	res := ""
 
@@ -33,4 +38,22 @@ func removeEmpty(s []string) []string {
 		}
 	}
 	return r
+}
+
+func readNonEmptyLines(path string) ([]string, error) {
+	file, err := os.Open(path)
+	if err != nil {
+		return nil, err
+	}
+	defer file.Close()
+
+	var lines []string
+	scanner := bufio.NewScanner(file)
+	scanner.Split(bufio.ScanLines)
+	for scanner.Scan() {
+		if scanner.Text() != "" {
+			lines = append(lines, scanner.Text())
+		}
+	}
+	return lines, scanner.Err()
 }