+++ helpers_test.go
@@ -12,7 +12,13 @@
 
 package tools4clj
 
-import "testing"
+import (
+	"bufio"
+	"fmt"
+	"os"
+	"strings"
+	"testing"
+)
 
 type TestJoinItem struct {
 	input    []string
@@ -61,5 +67,57 @@ func TestRemoveEmpty(t *testing.T) {
 				t.Errorf("remove empty slices failed, expected slice %v, got %v", v.expected[i], s)
 			}
 		}
+	}
+}
+
+func TestReadLines(t *testing.T) {
+	textLines := []string{
+		"This is a multiline text,",
+		"in order to test",
+		"read-lines function",
+		"not reading empty lines:",
+		"",
+		"not reading CR only lines:",
+		"\r",
+	}
+
+	text := strings.Join(textLines, "\n")
+
+	tempTextFile := "temptext.txt"
+	file, err := os.Create(tempTextFile)
+	if err != nil {
+		t.Errorf("failed to create test text file")
+	}
+	defer os.Remove(tempTextFile)
+
+	w := bufio.NewWriter(file)
+	_, err = fmt.Fprintln(w, text)
+	if err != nil {
+		t.Errorf("failed to write to test text file")
+	}
+	w.Flush()
+	err = file.Close()
+	if err != nil {
+		t.Errorf("failed to close test text file after writing to it")
+	}
+
+	lines, err := readNonEmptyLines(tempTextFile)
+	if err != nil {
+		t.Errorf("readlines error: %v", err)
+	}
+	i := 0
+	for _, line := range textLines {
+		if line == "" || line == "\r" {
+			continue
+		}
+		if lines[i] != line {
+			t.Errorf("read lines failed, expected \"%v\", got \"%v\"", line, lines[i])
+		}
+		i++
+	}
+
+	_, err = readNonEmptyLines("not_Existing_file.txt")
+	if err == nil {
+		t.Errorf("readlines error, reading non existing file")
 	}
 }